Square bins, rectangular plots.  Refs #1543.
All bins are now square, using the smaller of the bin width and height
as the bin size.  As a result, plots are now rectangular.

In the event that a plot is narrow, it is centered in the plot window.
In the event that a plot is wide, it is moved up in the plot window.