Disabled old prefilter, added is_hydrogen method.
The hydrogen check was changed to look for hydrogen or deuterium.  It was cleaner
to add it as its own method at this point.

Previously we prefiltered ATOM lines that had no valid amino acids and HETATM lines
with valid amino acids.  This prefiltering is no longer necessary and has an negative
impact on processing efficiency.